requirements.txt
Pillow == 8.2.0
PyYAML == 5.4.1
anytree == 2.8.0
matplotlib == 3.3.2
numpy == 1.20.1
pandas == 1.1.3
scipy == 1.6.2
six == 1.15.0
tabulate == 0.8.7
torch == 2.2.1
torchvision == 0.17.1
tqdm == 4.59.0
scikit-image
wandb
pyhessian
plotly
